--- services/app-web/src/pages/StateSubmission/RateDetails/V2/ReviewSubmit/ReviewSubmit.module.scss
@@ -1,13 +1,0 @@
-@use '../../../../../styles/custom.scss' as custom;
-@use '../../../../../styles/uswdsImports.scss' as uswds;
-
-.reviewSectionWrapper {
-    max-width: custom.$mcr-container-standard-width-fixed;
-}
-
-.submitButton {
-    background: custom.$mcr-success-base;
-    &:hover {
-        background-color: custom.$mcr-success-hover !important;
-    }
-}
