--- services/app-web/src/pages/StateSubmission/RateDetails/V2/ReviewSubmit/SubmissionSummarySection.module.scss
@@ -1,117 +1,0 @@
-@use '../../../../../styles/custom.scss' as custom;
-@use '../../../../../styles/uswdsImports.scss' as uswds;
-
-.summarySectionWrapper {
-    max-width: 50rem;
-}
-
-.summarySection {
-    margin: uswds.units(2) auto;
-    background: custom.$mcr-foundation-white;
-    padding: uswds.units(2) uswds.units(4);
-    border: 1px solid custom.$mcr-gray-lighter;
-    line-height: uswds.units(3);
-
-    h2 {
-        margin: 0;
-        @include uswds.u-text('normal');
-    }
-    &:first-of-type {
-        h2 {
-            @include uswds.u-text('bold');
-            font-size: size('body', 'lg');
-        }
-    }
-    ul {
-        list-style-type: none;
-        margin: 0;
-        padding: 0;
-
-        li {
-            padding: uswds.units(1) 0;
-        }
-    }
-
-    // align state contacts
-    dl div[class^='grid-container'] {
-        padding: 0;
-
-        div {
-            padding-bottom: uswds.units(2);
-        }
-
-        // dont bottom pad last two grid items
-        div:nth-last-child(-n + 2) {
-            padding-bottom: 0;
-        }
-    }
-    table:last-of-type {
-        margin-bottom: uswds.units(2);
-    }
-
-    // with nested sections, collapse bottom margin/padding for last in list
-    // rely on margin from the parent .summarySection. Relevant for multi-rate experience
-    section:last-of-type {
-        margin-bottom: 0 !important;
-    }
-}
-
-.contactInfo p {
-    margin: uswds.units(1) 0;
-}
-
-.documentDesc {
-    margin: 0;
-}
-
-.submitButton {
-    background: custom.$mcr-success-base;
-    &:hover {
-        background-color: custom.$mcr-success-hover !important;
-    }
-}
-
-.rateName {
-    display: block;
-    font-weight: bold;
-    margin: 0;
-}
-
-.certifyingActuaryDetail {
-    margin-bottom: uswds.units(4);
-
-    @include uswds.at-media(tablet) {
-        margin-bottom: 0;
-    }
-
-    dt {
-        font-weight: bold;
-    }
-
-    dd {
-        display: block;
-        margin: 0;
-        line-height: 1.5;
-    }
-}
-
-.singleColumnGrid {
-    @include uswds.at-media(tablet) {
-        > * {
-            margin-bottom: uswds.units(2);
-        }
-    }
-}
-
-@media print {
-    header,
-    footer,
-    .pageActions,
-    .summarySectionHeader a {
-        display: none;
-    }
-    .summarySection {
-        border: none;
-        padding: 0;
-    }
-}
