+++ services/app-web/src/pages/StateSubmission/StateSubmissionForm.tsx
@@ -21,7 +21,7 @@ import { UnlockedHealthPlanFormDataType } from '../../common-code/healthPlanForm
 import { useLDClient } from 'launchdarkly-react-client-sdk'
 import { featureFlags } from '../../common-code/featureFlags'
 import { RateDetailsV2 } from './RateDetails/V2/RateDetailsV2'
-import { ReviewSubmitV2 } from './RateDetails/V2/ReviewSubmit/ReviewSubmitV2'
+import { ReviewSubmitV2 } from './ReviewSubmit/V2/ReviewSubmit/ReviewSubmitV2'
 import styles from './StateSubmissionForm.module.scss'
 
 // Can move this AppRoutes on future pass - leaving it here now to make diff clear
